Require Auth. for reviewing requests
* Resolved circular imports
* Fixed test failures
* Fixed issue with DB connections in dynamic_routes.py
