+++ backend/__init__.py
@@ -6,6 +6,7 @@ from sqlalchemy import select
 from sqlalchemy.orm import Session, subqueryload
 
 from .config import settings
+from .crud import ALLOWED_FILTERS
 from .database import SessionLocal
 from .dynamic_routes import create_dynamic_router
 from .models import Schema, AttributeDefinition, BoundFK
@@ -30,7 +31,7 @@ def generate_api_description() -> str:
         description += '\n* `{}` - {}'.format(filter, desc)
 
     description += '\n\n**Available filters for each type**:'
-    for type, filters in crud.ALLOWED_FILTERS.items():
+    for type, filters in ALLOWED_FILTERS.items():
         description += '\n* `{}`: {}'.format(type.name, ', '.join([f'`{i}`' for i in filters]))
     return description
 
@@ -38,7 +39,7 @@ def generate_api_description() -> str:
 def load_dynamic_routes(db: Session, app: FastAPI):
     schemas = load_schemas(db)
     for schema in schemas:
-        create_dynamic_router(schema=schema, app=app, get_db=lambda: db)
+        create_dynamic_router(schema=schema, app=app)
 
 
 def create_app(session: Optional[Session] = None) -> FastAPI:
