+++ backend/alembic/versions/63c4cf462c86_review.py
@@ -1,0 +1,134 @@
+"""review
+
+Revision ID: 63c4cf462c86
+Revises: 5c72f33e6b82
+Create Date: 2022-01-25 09:13:12.543588
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '63c4cf462c86'
+down_revision = '5c72f33e6b82'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('change_values_bool',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('old_value', sa.Boolean(), nullable=True),
+    sa.Column('new_value', sa.Boolean(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_change_values_bool_id'), 'change_values_bool', ['id'], unique=False)
+    op.create_table('change_values_date',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('old_value', sa.Date(), nullable=True),
+    sa.Column('new_value', sa.Date(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_change_values_date_id'), 'change_values_date', ['id'], unique=False)
+    op.create_table('change_values_datetime',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('old_value', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('new_value', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_change_values_datetime_id'), 'change_values_datetime', ['id'], unique=False)
+    op.create_table('change_values_fk',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('old_value', sa.Integer(), nullable=True),
+    sa.Column('new_value', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_change_values_fk_id'), 'change_values_fk', ['id'], unique=False)
+    op.create_table('change_values_float',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('old_value', sa.Float(), nullable=True),
+    sa.Column('new_value', sa.Float(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_change_values_float_id'), 'change_values_float', ['id'], unique=False)
+    op.create_table('change_values_int',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('old_value', sa.Integer(), nullable=True),
+    sa.Column('new_value', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_change_values_int_id'), 'change_values_int', ['id'], unique=False)
+    op.create_table('change_values_str',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('old_value', sa.String(), nullable=True),
+    sa.Column('new_value', sa.String(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_change_values_str_id'), 'change_values_str', ['id'], unique=False)
+    op.create_table('change_requests',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('created_by_user_id', sa.Integer(), nullable=False),
+    sa.Column('reviewed_by_user_id', sa.Integer(), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('reviewed_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('status', sa.Enum('PENDING', 'DECLINED', 'APPROVED', name='changestatus'), nullable=False),
+    sa.Column('comment', sa.String(length=1024), nullable=True),
+    sa.Column('object_type', sa.Enum('SCHEMA', 'ENTITY', name='editableobjecttype'), nullable=False),
+    sa.Column('change_type', sa.Enum('CREATE', 'UPDATE', 'DELETE', name='changetype'), nullable=False),
+    sa.ForeignKeyConstraint(['created_by_user_id'], ['users.id'], ),
+    sa.ForeignKeyConstraint(['reviewed_by_user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('changes',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('change_request_id', sa.Integer(), nullable=False),
+    sa.Column('attribute_id', sa.Integer(), nullable=True),
+    sa.Column('object_id', sa.Integer(), nullable=True),
+    sa.Column('value_id', sa.Integer(), nullable=False),
+    sa.Column('content_type', sa.Enum('ATTRIBUTE', 'ATTRIBUTE_DEFINITION', 'ENTITY', 'SCHEMA', name='contenttype'), nullable=False),
+    sa.Column('change_type', sa.Enum('CREATE', 'UPDATE', 'DELETE', name='changetype'), nullable=False),
+    sa.Column('field_name', sa.String(), nullable=True),
+    sa.Column('data_type', sa.Enum('STR', 'BOOL', 'INT', 'FLOAT', 'FK', 'DT', 'DATE', name='changeattrtype'), nullable=False),
+    sa.CheckConstraint('NOT(attribute_id IS NULL AND field_name IS NULL)'),
+    sa.ForeignKeyConstraint(['attribute_id'], ['attributes.id'], ),
+    sa.ForeignKeyConstraint(['change_request_id'], ['change_requests.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('values_date',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('value', sa.Date(), nullable=True),
+    sa.Column('entity_id', sa.Integer(), nullable=True),
+    sa.Column('attribute_id', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['attribute_id'], ['attributes.id'], ),
+    sa.ForeignKeyConstraint(['entity_id'], ['entities.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_values_date_id'), 'values_date', ['id'], unique=False)
+    op.add_column('schemas', sa.Column('reviewable', sa.Boolean(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('schemas', 'reviewable')
+    op.drop_index(op.f('ix_values_date_id'), table_name='values_date')
+    op.drop_table('values_date')
+    op.drop_table('changes')
+    op.drop_table('change_requests')
+    op.drop_index(op.f('ix_change_values_str_id'), table_name='change_values_str')
+    op.drop_table('change_values_str')
+    op.drop_index(op.f('ix_change_values_int_id'), table_name='change_values_int')
+    op.drop_table('change_values_int')
+    op.drop_index(op.f('ix_change_values_float_id'), table_name='change_values_float')
+    op.drop_table('change_values_float')
+    op.drop_index(op.f('ix_change_values_fk_id'), table_name='change_values_fk')
+    op.drop_table('change_values_fk')
+    op.drop_index(op.f('ix_change_values_datetime_id'), table_name='change_values_datetime')
+    op.drop_table('change_values_datetime')
+    op.drop_index(op.f('ix_change_values_date_id'), table_name='change_values_date')
+    op.drop_table('change_values_date')
+    op.drop_index(op.f('ix_change_values_bool_id'), table_name='change_values_bool')
+    op.drop_table('change_values_bool')
+    # ### end Alembic commands ###
